Transfer Learning By Using Cats Vs Dogs Classification Problem
Transfer Learning

Transfer learning is one of the more important techniques of deep learning and tensorflow lets you do that with a handful of lines of codes. We needing to train a neural network from scratch we can need a lot of data and take a long time to train. We can instead download may be an open-source model that someone else has already trained on a huge dataset may be for weeks and us those parameters as a starting point to then train you model just a little bit more perhaps a smaller dataset that you have for given task . So it is called transfer learning.

Transfer learning so simple to implement in Tensorflow , we download a model and then we say set these model as trainable and freeze or lock those other layers and then we just run.

If train data very small and there’s only so many common feature that can be extracted even if we do some tricks like image augmentation. But in the both these cases we built the model from scratch.

Another useful tool to explore at this point is the Dropout.
The idea behind Dropouts is that they remove a random number of neurons in your neural network. This works very well for two reasons: The first is that neighboring neurons often end up with similar weights, which can lead to overfitting, so dropping some out at random can remove this. The second is that often a neuron can over-weigh the input from a neuron in the previous layer, and can over specialize as a result. Thus, dropping out can break the neural network out of this potential bad habit!
heck out Andrew's terrific video explaining dropouts here: https://www.youtube.com/watch?v=ARq74QuavAo